Criando novo estado de rotação
Novo estado de rotação, algumas mudanças, e tudo trocado de float pra
double

+++ TrabalhoCompGraf_1/beziercurve.cpp
@@ -13,20 +13,20 @@ BezierCurve::~BezierCurve(void) {
 void BezierCurve::refresh() {
 	m_curvePoints.clear();
 
-	for (float t = 0.0f; t <= 1.0f; t += 0.01f){
-		float x = 0.0f, y = 0.0f;
+	for (double t = 0.0f; t <= 1.0f; t += 0.01f){
+		double x = 0.0f, y = 0.0f;
 
 		for (size_t i = 0; i < getControlPoints().size(); i++) {
-			x += (float)(getControlPoints()[i].getX() * combination(getCurveDegree(), i) * pow(t, i) * pow( (1.0f - t), (getCurveDegree() - i)));
-			y += (float)(getControlPoints()[i].getY() * combination(getCurveDegree(), i) * pow(t, i) * pow( (1.0f - t), (getCurveDegree() - i)));
+			x += (double)(getControlPoints()[i].getX() * combination(getCurveDegree(), i) * pow(t, i) * pow( (1.0f - t), (getCurveDegree() - i)));
+			y += (double)(getControlPoints()[i].getY() * combination(getCurveDegree(), i) * pow(t, i) * pow( (1.0f - t), (getCurveDegree() - i)));
 		}
 		
 		addCurvePoint(x, y);
 	}
 }
 
-float BezierCurve::factorial(int value){
-	float fact = 1.0;
+double BezierCurve::factorial(int value){
+	double fact = 1.0;
 
 	for (size_t i = value; i > 0; i--)
 		fact *= i;
@@ -34,7 +34,7 @@ float BezierCurve::factorial(int value){
 	return fact;
 }
 
-float BezierCurve::combination(int n, int i){
+double BezierCurve::combination(int n, int i){
 	return factorial(n) / (factorial(i) * factorial(n-i));
 }
